requirements.txt
opencv-python-headless
pytesseract
pdf2image
pillow
loguru
numpy
